requirements.txt
Flask==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
gunicorn==20.1.0

